requirements.txt
fastapi
uvicorn[standard]
bcrypt
pydantic
greenlet
asyncpg
PyJWT
python-multipart